# ui/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Moongate.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/images/hair/{style}.png": {
      "get": {
        "tags": [
          "mobiles"
        ],
        "summary": "Serves a hair (or facial-hair) style as PNG, rendered over a reference body.",
        "description": "The style is the hair item's decimal graphic id (see /api/v1/admin/hair-styles). Pass hue to dye\nit, body to change the reference body (default 400, the human male), facial=true for beards.\n404 when nothing decodes for the pair.",
        "operationId": "GetHairImage",
        "parameters": [
          {
            "name": "style",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hue",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "body",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "facial",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/register": {
      "post": {
        "tags": [
          "registration"
        ],
        "summary": "Registers a new account when web registration is open.",
        "description": "Answers 403 when registration is disabled, 429 when the caller is rate-limited, 409 for a taken\nusername, 400 for a missing or malformed field, and 202 on success — the account is created\ninactive and must verify its email before it can log in.",
        "operationId": "RegisterAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/v1/register/verify": {
      "post": {
        "tags": [
          "registration"
        ],
        "summary": "Verifies an account's email with its token, activating the account.",
        "description": "Answers 400 for an unknown or already-used token.",
        "operationId": "VerifyRegistration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/server-info/assets/{slot}": {
      "get": {
        "tags": [
          "server-info"
        ],
        "summary": "Streams a visual asset (logo, favicon or banner) by slot.",
        "description": "Answers 400 for an unknown slot and 404 when the slot has no asset.",
        "operationId": "GetServerAsset",
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/server-settings": {
      "get": {
        "tags": [
          "server-settings"
        ],
        "summary": "Returns the full server settings.",
        "operationId": "GetServerSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "server-settings"
        ],
        "summary": "Updates the server settings; every field is optional and an omitted one is left unchanged.",
        "operationId": "UpdateServerSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServerSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/server-settings/assets/{slot}": {
      "post": {
        "tags": [
          "server-settings"
        ],
        "summary": "Uploads the image for a slot (logo, favicon or banner), replacing any previous one.",
        "description": "Answers 400 for an unknown slot, 415 for a non-image type, and 413 when the file is too large.",
        "operationId": "UploadServerAsset",
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerSettingsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "server-settings"
        ],
        "summary": "Removes the image for a slot.",
        "operationId": "DeleteServerAsset",
        "parameters": [
          {
            "name": "slot",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "A web self-registration request."
      },
      "ServerContactsResponse": {
        "type": "object",
        "properties": {
          "website": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "discord": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The shard's public contact points."
      },
...
      "ServerSettingsResponse": {
        "required": [
          "assets",
          "contacts",
          "registrationEnabled"
        ],
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "contacts": {
            "$ref": "#/components/schemas/ServerContactsResponse"
          },
          "registrationEnabled": {
            "type": "boolean"
          },
          "assets": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false,
        "description": "The full settings view returned to staff (identical fields to the public info minus the shard name)."
      },
...
      "UpdateServerSettingsRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "contacts": {
            "$ref": "#/components/schemas/ServerContactsResponse"
          },
          "registrationEnabled": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "A partial settings update: an omitted (null) field is left unchanged."
      },
      "VerifyEmailRequest": {
        "required": [
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "Carries the single-use token that verifies an account's email."
      },
...
    }
  },
...
}